JOPEINDWRK-59 Vending Machine COMPLETE with flow chart and class diagrams

+++ week03/VendingMachine/src/main/java/com/jorypestorious/vendingmachine/ui/ConsoleIO.java
@@ -50,7 +50,7 @@ public class ConsoleIO {
                 scan.next();
             }
             
-            input = scan.nextInt();
+            input = scan.nextDouble();
             
             if (input > max || input < min) {
                 System.out.print(errorMsg);
